+++ headers_cpp/extend_streamlines.h
@@ -6,12 +6,12 @@
 #include "streamlines.h"
 
 
-std::vector<Point> extend_streamline(const Matrix& V, const std::vector<Point>& streamline, float step=0.5f, int max_length=1000);
+std::vector<Point> extend_streamline(const Matrix& V, const std::vector<Point>& streamline, double step=0.5, int max_length=1000);
 
 
 std::vector<std::vector<Point>> get_extended_streamlines(   const Matrix& B, const Matrix& V,
                                                             const std::vector<Point>& points, 
-                                                            float step=0.5f, int max_length=1000,
+                                                            double step=0.5, int max_length=1000,
                                                             const Point* earth_pos=nullptr  );
