+++ src_cpp/extend_streamlines.cpp
@@ -10,9 +10,9 @@
 
 
 
-std::vector<Point> extend_streamline(const Matrix& V, const std::vector<Point>& streamline, float step, int max_length)
+std::vector<Point> extend_streamline(const Matrix& V, const std::vector<Point>& streamline, double step, int max_length)
 {
-    float starting_x = V.get_shape().x * 0.8;         // TODO: FIX MAGIC NUMBER
+    double starting_x = V.get_shape().x * 0.8;         // TODO: FIX MAGIC NUMBER
     std::vector<Point> extended_streamline;
 
     for (int i=(int) streamline.size()-1; i>=0; i--)
@@ -40,7 +40,7 @@ std::vector<Point> extend_streamline(const Matrix& V, const std::vector<Point>& 
 
 std::vector<std::vector<Point>> get_extended_streamlines(   const Matrix& B, const Matrix& V,
                                                             const std::vector<Point>& points, 
-                                                            float step, int max_length,
+                                                            double step, int max_length,
                                                             const Point* earth_pos  )
 {
     std::vector<std::vector<Point>> streamlines = get_streamlines(B, points, step, max_length, earth_pos);
